Added a bunch
ICommand now has an Admin property. CommandRouter will check this and not execute if the executing user isn't a controller. Saves trying to implement the logic yourself.

Also added the config command, which allows saving the config, and changing admins.

Also the config.json location can be specified by CLA 3 so it doesn't get overwritten when the app gets updated.